pyproject.toml
[tool.poetry]
name = "the_wall"
version = "0.1.0"
description = "Calculating material cost and quantities for building The Wall(tm)"
authors = ["VSh <vl.shumanov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
